Added type-bound procedures mean and interpolate for tPolicies
And also made allocate_policies depend on nz, so that I could create
policies for a specific z (e.g. in mean shock or in simulations).
However, while they technically work (I believe), they are not useful at
the moment, because they result in a type tPolicies, while I typically
want the result to have one dimension less. This is not easy to
implement, though.