Made Euler equation internal procedure
Moved asseteuler_f from asseteuler to policyfunctions, thereby making it
an internal procedure of solve_policyfunctions. The passing of
internal procedures as arguments is Fortran 2008 standard, but supported
by gfortran and Intel since a while. It should be thread safe (next
step). This allowed me to make the whole solve_policyfunctions pure.