Introduced new variables in policyfunctions
and assigned them to the derived type allocatable components. Then I
made many THREADPRIVATE, because that is some kind of global (hoping
that this makes them visible to the contained subroutines). For
THREADPRIVATE, the variables need to be SAVEd, and for the latter, I had
to make them allocatable instad of automatic arrays...

The program compiles and runs but segfaults pretty soon...

Probably I could reduce the list of variables that are THREADPRIVATE.