Initial branch commit with plan
Want to put meanshock_equilib into meanshock_wrapper. That way, it is
analogous to krusell_smith, and also to the root-finding code structure
in household_solution_mod:asset_allocation. I do not want to write the
additional results directly into the host variables, but instead have a
subroutine that mirrors the function but returns more arguments. This
replaces the getter method, and has the additional advantage that the
whole meanshock thing can become pure, and that I do not
interpolateXgrid in the function (only in the subroutine).

In a second step, I might call simulate_economy from here also.