Initial branch 020 commit with plan
Plan: Want to update aggregate grids in krusell_smith so that we don't
hit the boundaries so often. Introduce a logical update_aggr_grids.
Probably don't wanna remove the factors that scale up the ms solution.
The question is whether I should update the grid only once in the
beginning before starting the rootfinder, or every x runs, etc...