Moved grid update and calc of period 1 simvars from KS to run_model
The problem was that I had previously changed krusell_smith so that the
first run previous to the rootfinder was not performed anymore. In
principle, that was correct, because it had implied a full Newton step
on the coefficients, which was bad as then it could happen that at the
start of the rootfinder KS wouldn't converge. However, due to this
change, the simvars were not defined, so that calculating the statistics
of simvars led to senseless results, and consequently the period 1
simvars were senseless.

Now, these steps are done in run_model_mod instead