Save coeffs of loms that lead to last solution, not those after
So far, the coeffs in the last iteration were updated after the
regression, then the Newton for KS was stopped. However, that means that
we do not have 1-to-1 the coeffs that lead to the solution; indeed the
last update implies a full Newton step which might be bad. Now, the loms
that are input into the last iteration before convergence are saved.